[QCF-69] 좋아요 API Tabs -> boardItem으로 이동
