+++ desk/src/components/units/user/components/boardItem/boardItem.queries.ts
@@ -1,0 +1,7 @@
+import { gql } from '@apollo/client'
+
+export const UPDATE_BOARD_LIKER = gql`
+  mutation updateBoardLiker($boardid: String!) {
+    updateBoardLiker(boardid: $boardid)
+  }
+`
