+++ desk/src/components/units/user/components/tabs/index.tsx
@@ -23,7 +23,6 @@ import {
   FETCH_BOARDS_USER_LIKED,
   FETCH_PRODUCTS,
   FETCH_USER_BOARDS,
-  UPDATE_BOARD_LIKER,
 } from './Tabs.queries'
 import {
   TMutation,
@@ -56,11 +55,6 @@ export default function NavigationTabs(props: NavigationTabsProps) {
   const { data: userProducts } = useQuery<Pick<TQuery, 'fetchProducts'>>(FETCH_PRODUCTS, {
     variables: { userid: props.userid as string },
   })
-
-  const [updateBoardLiker] = useMutation<
-    Pick<TMutation, 'updateBoardLiker'>,
-    TMutationUpdateBoardLikerArgs
-  >(UPDATE_BOARD_LIKER)
 
   const TABS = props.isMyPage ? MY_PAGE_TAB : OTHERS_PAGE_TAB
 
@@ -105,14 +99,6 @@ export default function NavigationTabs(props: NavigationTabsProps) {
     [showUserPosts, showUserProductPosts, showLikedPosts],
   )
 
-  const onClickLikeButton = async (boardid: string) => {
-    await updateBoardLiker({
-      variables: {
-        boardid,
-      },
-    })
-  }
-
   return (
     <>
       <Tabs mt="auto">
@@ -150,7 +136,6 @@ export default function NavigationTabs(props: NavigationTabsProps) {
                   boardId={item.id}
                   imageUrl={item.pictures[0].url}
                   isLiked={item.like}
-                  onClickLikeButton={() => onClickLikeButton(item.id)}
                 />
               )}
             </React.Fragment>
